Added hmac functions to common file
Added 3 functions to common file -
1) get_HMAC
2) verify HMAC
3) verify and asymmetric decrypt - This functions can be used on the
incoming message which was encrypted with AES and signed with HMAC. It
verifies hmac and does the symmetric decryption as well and returns
plain text.

Should add another function with symmetrically encrypts and adds HMAC in
one go.

Added a HMAC verification function to test file.